New: interface with udev database
Rather than attempting to parse `/dev` and `/sys` directly within smfc,
we can outsource that task to `udev` (through [pyudev][1]). This allows
us to simplify much of the logic and string parsing, making it less
fragile and more robust.

Unfortunately, the unit tests will require a more extensive refactoring
to be able to support pyudev, but these changes are not a part of this
commit. If we want to keep the current strategy of creating a set of
"test data" for the tests, then something like [umockdev][2] might be
best. Otherwise, we could write `Mock`s for the expected function calls,
but for that a wholesale restructuring of all the test data is probably
required.

[1]: http://pyudev.readthedocs.org/
[2]: https://github.com/martinpitt/umockdev