Fix: NVMe path detection
Under linux (under the persistent storage naming rules present in most
distributions AFAIK), the NVMe device by-id symlinks point to the block
device (e.g., `/dev/nvme0n1`) rather than the controller (`/dev/nvme0`).
However, the corresponding `hwmon` device is actually a child of the
controller (thus a sibling to the block device in sysfs).

The original code has a bug where the symlink is assumed to be backwards
(so it assumes the link points to `/dev/nvme0` and concatenates the
namespace within sysfs). This bug is also present in the unit tests in
the test data created by the file `test_00_data.py`.

Properly parsing the string would require stripping the namespace (in
99% of cases, this is `"n1"`) from the block device name. Instead, in
this commit we use pyudev to query the udev database to find the sibling
`hwmon` device. Unfortunately, the test file doesn't have such a simple
fix, so someone will have to patch it more deeply.